Move abi to dedicated module
