+++ crates/client/rpc-core/src/lib.rs
@@ -42,6 +42,7 @@ pub trait StarknetRpcApi {
     /// address
     #[method(name = "getClassHashAt")]
     fn get_class_hash_at(&self, contract_address: ContractAddress, block_id: BlockId) -> RpcResult<FieldElement>;
+
     /// Get an object about the sync status, or false if the node is not syncing
     #[method(name = "syncing")]
     async fn syncing(&self) -> RpcResult<Syncing>;
