+++ crates/client/rpc-core/src/types/mod.rs
@@ -1,41 +1,26 @@
+mod abi;
 mod block;
-mod contract_abi;
 mod deprecated;
-mod events;
-mod function_abi;
 mod function_call;
 mod rpc_contract_class;
 mod sierra_contract_class;
-mod struct_abi;
 mod syncing;
 mod typed_parameter;
 
-use std::collections::HashMap;
-
+pub use abi::*;
 pub use block::*;
-pub use contract_abi::*;
 pub use deprecated::*;
-pub use events::*;
-pub use function_abi::*;
-pub use function_call::FunctionCall;
-pub use rpc_contract_class::RPCContractClass;
+pub use function_call::*;
+pub use rpc_contract_class::*;
 pub use sierra_contract_class::*;
-pub use struct_abi::*;
 pub use syncing::*;
-pub use typed_parameter::TypedParameter;
+pub use typed_parameter::*;
 
 pub type ContractAddress = FieldElement;
 pub type ContractClassHash = FieldElement;
+pub type ContractClassVersion = String;
 pub type FieldElement = String;
-pub type FunctionName = String;
-pub type Members = Vec<StructMember>;
 pub type Offset = String;
 pub type Program = String;
 pub type Selector = FieldElement;
-pub type Size = i64;
-pub type StructMember = HashMap<String, serde_json::Value>;
-pub type StructName = String;
 pub type SierraProgram = Vec<FieldElement>;
-pub type ContractClassVersion = String;
-pub type FunctionIndex = i64;
-pub type ABI = String;
